Require directory when launching applications.
Applications are no longer assumed to be in the 'public/apps' directory.
When launching the vroom.js server the path relative to the 'public/'
directory is used. For example, instead of running
   vroom-server.js simple.js
you should call
   vroom-server.js apps/simple.js
This makes it possible to launch applications (for live-coding) from
any sub-directory (apps, examples, etc).